feat: Se añadio estilos al formulario del cliente y algunas correcciones
