+++ frent/src/components/Form/customerForm.css
@@ -1,0 +1,101 @@
+@import url('https://fonts.googleapis.com/css2?family=Merriweather:ital,wght@0,300;0,400;0,700;0,900;1,300;1,400;1,700;1,900&display=swap');
+@import url('https://fonts.googleapis.com/css2?family=Merriweather:ital,wght@0,300;0,400;0,700;0,900;1,300;1,400;1,700;1,900&family=Poppins:ital,wght@0,100;0,200;0,300;0,400;0,500;0,600;0,700;0,800;0,900;1,100;1,200;1,300;1,400;1,500;1,600;1,700;1,800;1,900&display=swap');
+
+.FormularioRegistro{
+  display: flex;
+  justify-content: center;
+  padding: 30px 0px;
+  background-color: #f0f0f0;
+}
+.container-principal{
+  background: #fff;
+  width: 50%;
+  border-radius: 12px;
+  font-family: "Poppins", sans-serif;
+  font-weight: 400;
+  font-style: normal;
+  box-shadow: 0px 4px 8px rgba(0, 0, 0, 0.1);
+}
+
+input {
+  border: 2px solid #ccc;
+  border-radius: 4px;
+}
+
+input::placeholder{
+  color: #7b7b7b;
+}
+
+select {
+  border: 2px solid #ccc;
+  border-radius: 4px; 
+}
+
+#formulario-cliente {
+    display: flex;
+    justify-content: center;
+  }
+  
+  #formulario-cliente label {
+    margin-bottom: 5px;
+  }
+  
+  .container-principal > h2 {
+    text-align: center;
+  }
+  
+  .container {
+    display: flex;
+    justify-content: center;
+    flex-wrap: wrap;
+  }
+  
+  .campos-form {
+    display: flex;
+    flex-direction: column;
+  }
+  
+  .campos-mitad {
+    width: 42%;
+    margin: 10px 10px;
+  }
+  
+  .campos-mitad-mitad {
+    width: 20%;
+    margin: 10px 10px;
+  }
+  
+  .campo-full {
+    width: 86%;
+  }
+  
+  input,
+  select {
+    border-radius: 5px;
+    padding: 10px;
+  }
+  
+  .llevar-inicio {
+    margin-right: 67%;
+  }
+  
+  .buttons-cancel-register {
+    width: 86%;
+    display: flex;
+    justify-content: space-between;
+    padding: 20px;
+  }
+  
+  .buttons-cancel-register > button {
+    font-size: 16px;
+    padding: 12px 18px;
+    border-radius: 8px;
+  }
+
+  .RecordButton{
+    background-color: #FBA834;
+  }
+  
+
+
+
